Setup router
