+++ order-portal/src/App.tsx
@@ -1,23 +1,22 @@
-import { Container, Heading, Box, Button } from '@chakra-ui/react';
 import React from 'react';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import OrderCreate from './pages/OrderCreate';
+import OrderIndex from './pages/OrderIndex';
 
 function App() {
-  const orders = ['abc', 'def', 'ghi', 'klm'];
 
   return (
     <div className="App">
-      <Container>
-        <Heading as="h1" textAlign="center" py="4">Orders</Heading>
-        
-        <Box display="flex" justifyContent="end">
-          <Button>Create</Button>
-        </Box>
-
-        {orders.map((order) => (
-          <Box key={order}>{order}</Box>
-        ))}
-
-      </Container>
+      <Router>
+        <Switch>
+          <Route path="/create">
+            <OrderCreate />
+          </Route>
+          <Route path="">
+            <OrderIndex />
+          </Route>
+        </Switch>
+      </Router>
     </div>
   );
 }
